Successfully implemented Google Place Autocomplete
