+++ src/app/layout.tsx
@@ -3,6 +3,7 @@ import { Poppins } from 'next/font/google';
 import { ClerkProvider } from '@clerk/nextjs';
 
 import './globals.css';
+import { APIProvider } from '@vis.gl/react-google-maps';
 
 const poppins = Poppins({
   subsets: ['latin'],
@@ -25,9 +26,11 @@ export default function RootLayout({
 }) {
   return (
     <ClerkProvider>
-      <html lang='en'>
-        <body className={poppins.variable}>{children}</body>
-      </html>
+      <APIProvider apiKey={process.env.NEXT_PUBLIC_GOOGLE_MAPS_API_KEY}>
+        <html lang='en'>
+          <body className={poppins.variable}>{children}</body>
+        </html>
+      </APIProvider>
     </ClerkProvider>
   );
 }
